Fix morphological filter doctest failures
- Fix voxelize_binary_mask dtype issue by preserving original dtype instead of casting to float
- Update image_dilate_erode doctest to use the non-deprecated close() method
- Both fixes maintain the same functionality while resolving doctest failures

+++ pyvista/core/filters/data_set.py
@@ -7869,7 +7869,7 @@ class DataSetFilters(_BoundsSizeMixin, DataObjectFilters):
             if background_value == 0
             else np.ones(scalars_shape, dtype=scalars_dtype) * background_value
         )
-        binary_mask['mask'] = scalars.astype(float)
+        binary_mask['mask'] = scalars
         # Make sure that we have a clean triangle-strip polydata
         # Note: Poly was partially pre-processed earlier
         poly_ijk = poly_ijk.strip()
